Convert to monorepo for future vscode extension integration

+++ packages/cli/scripts/preuninstall.js
@@ -7,6 +7,9 @@ let imperativeSettings = {
 };
 if (fs.existsSync(settingsJson)) {
     imperativeSettings = JSON.parse(fs.readFileSync(settingsJson, "utf-8"));
+    imperativeSettings.overrides.CredentialManager = "@zowe/cli";
 }
-imperativeSettings.overrides.CredentialManager = "@zowe/cli";
-fs.writeFileSync(settingsJson, JSON.stringify(imperativeSettings, null, 2));
+
+if(fs.existsSync(settingsJson)) {
+  fs.writeFileSync(settingsJson, JSON.stringify(imperativeSettings, null, 2));
+}
